# docs/openapi.yaml
openapi: 3.2.0
info:
  title: kpblog API
  version: 0.1.0
  description: |
    kpblog 后端 JSON API 契约。

    所有 `/api/*` JSON 接口统一使用 envelope:

    - 成功: `{ "success": true, "data": ... }`
    - 失败: `{ "success": false, "error": { "code": "...", "message": "..." } }`

    Web 浏览器通过同域 `/api/*` 访问；Next.js 将请求代理到独立 Fastify
    进程。`/rss.xml` 等由 Web 输出的非 JSON 内容不包含在本文件中。
servers:
  - url: http://localhost:3001
    description: Local same-origin Web proxy
  - url: http://127.0.0.1:3002
    description: Local Fastify API (private)
  - url: https://kpblog.cc
    description: Production
...
paths:
...
  /api/public/guestbook:
    get:
      tags: [Public]
      summary: List approved guestbook entries
      description: >-
        Guestbook entries are stored in the same table as article comments with a
        null postId, so they share the review pipeline, one-level replies and
        rate limiting. Top-level entries are paginated, newest first (default 20
        per page, maximum 50); each entry carries its approved replies in full.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Approved guestbook threads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
...
  /api/public/article-index:
    get:
      tags: [Public]
      summary: Get the server-rendered article index payload
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
          description: >-
            Keyword filter over title, excerpt and content of published
            articles. The API truncates the value to 100 characters.
        - name: category
          in: query
          schema:
            type: string
          description: Category slug filter (the /articles page filter dropdown).
        - name: tag
          in: query
          schema:
            type: string
          description: Tag slug filter (the /articles page filter dropdown).
      responses:
        "200":
          description: Paginated published articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
...
  /api/public/articles/{slug}/adjacent:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get previous, next, and related articles for a published article
      responses:
        "200":
          description: Adjacent navigation payload, or null when the slug is unknown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleAdjacentResponse"
  /api/public/tags/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get a tag archive page payload
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Tag metadata and paginated articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/public/categories/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get a category archive page payload
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Category metadata and paginated articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/public/collections/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get a project collection page payload
      description: Server-rendered /collections/[collection] page data — the project summary plus a paginated list of its published posts.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Project page data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPageResponse"
...
  /api/music:
    get:
      tags: [Media]
      summary: List music tracks
      responses:
        "200":
          description: Music tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicListResponse"
    post:
      tags: [Media]
      summary: Upload local music file or add external music URL
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MusicUploadInput"
          application/json:
            schema:
              $ref: "#/components/schemas/MusicUrlInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Music track created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/music/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    delete:
      tags: [Media]
      summary: Delete music track
      security:
        - sessionCookie: []
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Music track deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/images:
    get:
      tags: [Media]
      summary: List registered images
      description: >-
        Admin only. The image library may contain unpublished article assets,
        private files, or external URLs with signed query parameters; the public
        site does not read from it. kind selects one bucket (cover or article);
        omit it for both. An unknown kind is rejected with 400.
      security:
        - sessionCookie: []
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [cover, article]
          description: Filter by image kind; omitted means all kinds.
      responses:
        "200":
          description: Registered images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Media]
      summary: Upload an image file or register an external image URL
      description: >-
        Admin only. Multipart uploads store the file under the local image
        library; JSON registers an external http(s) URL (e.g. a GitHub image
        host) without downloading it. Both create a MediaImage entry whose
        url must be unique.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUploadInput"
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUrlRegisterInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Image registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/images/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    delete:
      tags: [Media]
      summary: Delete a registered image
      description: |
        Admin only. Refuses with 409 while posts still reference the URL
        (covers match Post.coverImage exactly; article images match the body
        markdown). Supplying force=true deletes anyway, leaving the dangling
        references in place — posts are never modified.
      security:
        - sessionCookie: []
      parameters:
        - name: force
          in: query
          required: false
          schema:
            type: string
          description: '"true" deletes even while references remain.'
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Still referenced by posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailure"
  /api/images/adopt:
    post:
      tags: [Media]
      summary: Adopt image URLs already used by posts into the library
      description: |
        Admin only. Scans every post: coverImage values become cover entries,
        markdown body image URLs become article entries. Idempotent — URLs
        already registered are skipped, so a repeated run reports zero.
      security:
        - sessionCookie: []
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Adoption result (newly registered counts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageAdoptResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/wallpapers:
    get:
      tags: [Media]
      summary: List homepage wallpapers
      description: |
        Public list returns only enabled wallpapers in rotation order. With an
        authenticated admin session, `all=true` returns every entry including
        disabled ones. First access auto-seeds the repository's default
        wallpapers.
      parameters:
        - name: all
          in: query
          required: false
          schema:
            type: string
          description: '"true" with an admin session includes disabled entries.'
      responses:
        "200":
          description: Wallpapers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallpaperListResponse"
    post:
      tags: [Media]
      summary: Upload a homepage wallpaper (JPG/PNG/WebP/GIF/AVIF, max 10MB)
      description: Admin only. Appends the wallpaper to the end of the rotation, enabled.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/WallpaperUploadInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Wallpaper added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallpaperResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/wallpapers/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      tags: [Media]
      summary: Enable or disable a wallpaper
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WallpaperUpdateInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Wallpaper updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallpaperResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Media]
      summary: Remove a wallpaper from the rotation
      description: Admin only. Removes the rotation entry only; the underlying file is kept.
      security:
        - sessionCookie: []
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Wallpaper removed from rotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/wallpapers/reorder:
    post:
      tags: [Media]
      summary: Reorder homepage wallpapers
      description: Admin only. Ids are assigned rotation order in the order given; unlisted entries keep their relative order at the end.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WallpaperReorderInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Reordered (full admin list)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallpaperListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: kp API key
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
  schemas:
    ApiFailure:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, TOO_MANY_REQUESTS, INTERNAL_ERROR]
        message:
          type: string
        retryAfterSeconds:
          type: integer
          minimum: 1
          description: >-
            Only present on rate-limit (429) responses: seconds until the caller may
            try again. An ADMIN account allows exactly 3 password attempts: attempts
            1-3 return 401, the 4th is refused before password verification and locks
            the account for 15 minutes (configurable via ADMIN_LOGIN_LOCKOUT_MINUTES).
            Regular accounts allow 10 failures per 15 minutes. The lock follows the
            account, not the IP, so rotating IPs does not clear it.
...
    Envelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data: {}
    DeleteResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [deleted]
              properties:
                deleted:
                  type: boolean
                  const: true
...
    TagRef:
      type: object
      required: [name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    CategoryRef:
      type:
        - object
        - "null"
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
    PostSummary:
      type: object
      required: [id, slug, title, excerpt, coverImage, published, publishedAt, createdAt, updatedAt, category, tags]
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        excerpt:
          type:
            - string
            - "null"
        coverImage:
          type:
            - string
            - "null"
        published:
          type: boolean
        publishedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/CategoryRef"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRef"
...
    AdjacentRef:
      type:
        - object
        - "null"
      required: [slug, title]
      properties:
        slug:
          type: string
        title:
          type: string
...
    ArticleAdjacentResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type:
                - object
                - "null"
              required: [previous, next, related]
              properties:
                previous:
                  $ref: "#/components/schemas/AdjacentRef"
                next:
                  $ref: "#/components/schemas/AdjacentRef"
                related:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostSummary"
    ArticleListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page, pageSize, totalPages]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostSummary"
                total:
                  type: integer
                page:
                  type: integer
                pageSize:
                  type: integer
                totalPages:
                  type: integer
...
    CommentReply:
      type: object
      required: [id, author, content, createdAt]
      properties:
        id:
          type: string
        author:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentThread:
      allOf:
        - $ref: "#/components/schemas/CommentReply"
        - type: object
          required: [replies]
          properties:
            replies:
              type: array
              items:
                $ref: "#/components/schemas/CommentReply"
    CommentListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page, pageSize, totalPages]
              description: >-
                Top-level comment threads are paginated. Replies are returned in
                full inside each thread and are not paginated separately.
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/CommentThread"
                total:
                  type: integer
                page:
                  type: integer
                pageSize:
                  type: integer
                totalPages:
                  type: integer
...
    CollectionPageResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [project, articles]
              properties:
                project:
                  type:
                    - object
                    - "null"
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    slug:
                      type: string
                    description:
                      type: string
                    coverImage:
                      type:
                        - string
                        - "null"
                articles:
                  $ref: "#/components/schemas/ArticleListPayload"
    ArticleListPayload:
      type: object
      required: [items, total, page, pageSize, totalPages]
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PostSummary"
    WallpaperItem:
      type: object
      required: [id, url, enabled, sortOrder, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
        enabled:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    WallpaperListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/WallpaperItem"
    WallpaperResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/WallpaperItem"
    WallpaperUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    WallpaperUpdateInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    WallpaperReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          description: Wallpaper ids in the desired rotation order.
    ImageUploadInput:
      type: object
      required: [file, kind]
      properties:
        file:
          type: string
          format: binary
        kind:
          type: string
          enum: [cover, article]
    ImageUrlRegisterInput:
      type: object
      required: [url, kind]
      properties:
        url:
          type: string
          maxLength: 2048
          description: http(s) image URL from an external image host.
        kind:
          type: string
          enum: [cover, article]
        name:
          type: string
          description: Optional display name; defaults to the URL's last segment.
    ImageItem:
      type: object
      required: [id, kind, name, url, createdAt]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [cover, article]
        name:
          type: string
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    ImageResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImageItem"
    ImageAdoptResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [covers, articles]
              properties:
                covers:
                  type: integer
                  description: Newly registered cover entries.
                articles:
                  type: integer
                  description: Newly registered article-image entries.
    ImageListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ImageItem"
    MusicTrack:
      type: object
      required: [id, title, artist, url, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type:
            - string
            - "null"
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    MusicListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/MusicTrack"
    MusicResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MusicTrack"
    MusicUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        title:
          type: string
        artist:
          type: string
    MusicUrlInput:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        artist:
          type:
            - string
            - "null"
        url:
          type: string
...